@@ -1,10 +1,10 @@
 name: shutdown
 description: Prioritized events for graceful shutdown in server Dart applications.
-version: 0.4.2
+version: 0.4.3
 homepage: https://github.com/isoos/shutdown
 
 environment:
-  sdk: '>=2.7.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 #dependencies:
 #  path: ^1.4.1
